Re-update wallinObj

+++ src/objectives/wallinObjective.cpp
@@ -161,7 +161,7 @@ namespace ghost
     bestCost = v_cost( vecVariables, domain );
     double currentCost = bestCost;
 
-    int postprocessTimeLimit = std::max( 1, static_cast<int>( ceil(OPT_TIME / 100) ) );
+    int postprocessTimeLimit = std::max( 1, static_cast<int>( ceil( static_cast<double>(OPT_TIME) / 100) ) );
 
     while( (postprocesstimer = chrono::system_clock::now() - startPostprocess).count() < postprocessTimeLimit && bestCost > 0 )
     {
